Refactor generateTypes
