+++ packages/manager/src/features/Domains/DomainDetail/DomainRecords/DomainRecords.tsx
@@ -1,7 +1,7 @@
 import { deleteDomainRecord as deleteDomain } from '@linode/api-v4/lib/domains';
-import { Button, Typography } from '@linode/ui';
+import { Typography } from '@linode/ui';
 import Grid from '@mui/material/Unstable_Grid2';
-import { compose, isEmpty, lensPath, over, pathOr } from 'ramda';
+import { lensPath, over } from 'ramda';
 import * as React from 'react';
 
 import { ActionsPanel } from 'src/components/ActionsPanel/ActionsPanel';
@@ -15,19 +15,13 @@ import {
 } from 'src/utilities/errorUtils';
 import { scrollErrorIntoView } from 'src/utilities/scrollErrorIntoView';
 import { storage } from 'src/utilities/storage';
-import { truncateEnd } from 'src/utilities/truncate';
 
-import { DomainRecordActionMenu } from './DomainRecordActionMenu';
 import { DomainRecordDrawer } from './DomainRecordDrawer';
 import { StyledDiv, StyledGrid } from './DomainRecords.styles';
-import {
-  getNSRecords,
-  getTTL,
-  msToReadable,
-  typeEq,
-} from './DomainRecordsUtils';
 import { DomainRecordTable } from './DomainRecordTable';
+import { generateTypes } from './generateTypes';
 
+import type { GenerateTypesHandlers, IType } from './generateTypes';
 import type {
   Domain,
   DomainRecord,
@@ -68,395 +62,8 @@ interface State {
   types: IType[];
 }
 
-export interface IType {
-  columns: {
-    render: (r: Domain | DomainRecord) => JSX.Element | null | string;
-    title: string;
-  }[];
-  data: any[];
-  link?: () => JSX.Element | null;
-  order: 'asc' | 'desc';
-  orderBy: 'domain' | 'name' | 'target';
-  title: string;
-}
-
-const createLink = (title: string, handler: () => void) => (
-  <Button buttonType="primary" className="domain-btn" onClick={handler}>
-    {title}
-  </Button>
-);
-
 export const DomainRecords = (props: Props) => {
   const { domain, domainRecords, updateDomain, updateRecords } = props;
-
-  const generateTypes = (): IType[] => [
-    /** SOA Record */
-    {
-      columns: [
-        {
-          render: (d: Domain) => d.domain,
-          title: 'Primary Domain',
-        },
-        {
-          render: (d: Domain) => d.soa_email,
-          title: 'Email',
-        },
-        {
-          render: getTTL,
-          title: 'Default TTL',
-        },
-        {
-          render: compose(msToReadable, pathOr(0, ['refresh_sec'])),
-          title: 'Refresh Rate',
-        },
-        {
-          render: compose(msToReadable, pathOr(0, ['retry_sec'])),
-          title: 'Retry Rate',
-        },
-        {
-          render: compose(msToReadable, pathOr(0, ['expire_sec'])),
-          title: 'Expire Time',
-        },
-        {
-          render: (d: Domain) => {
-            return d.type === 'master' ? (
-              <DomainRecordActionMenu
-                editPayload={d}
-                label={domain.domain}
-                onEdit={handleOpenSOADrawer}
-              />
-            ) : null;
-          },
-          title: '',
-        },
-      ],
-      data: [domain],
-      order: 'asc',
-      orderBy: 'domain',
-      title: 'SOA Record',
-    },
-
-    /** NS Record */
-    {
-      columns: [
-        {
-          render: (r: DomainRecord) => r.target,
-          title: 'Name Server',
-        },
-        {
-          render: (r: DomainRecord) => {
-            const sd = r.name;
-            return isEmpty(sd) ? domain.domain : `${sd}.${domain.domain}`;
-          },
-          title: 'Subdomain',
-        },
-        {
-          render: getTTL,
-          title: 'TTL',
-        },
-        {
-          /**
-           * If the NS is one of Linode's, don't display the Action menu since the user
-           * cannot make changes to Linode's nameservers.
-           */
-          render: (domainRecordParams: DomainRecord) => {
-            const { id, name, target, ttl_sec } = domainRecordParams;
-
-            if (id === -1) {
-              return null;
-            }
-
-            return (
-              <DomainRecordActionMenu
-                deleteData={{
-                  onDelete: confirmDeletion,
-                  recordID: id,
-                }}
-                editPayload={{
-                  id,
-                  name,
-                  target,
-                  ttl_sec,
-                }}
-                label={name}
-                onEdit={openForEditNSRecord}
-              />
-            );
-          },
-          title: '',
-        },
-      ],
-      data: getNSRecords(props),
-      link: () => createLink('Add an NS Record', openForCreateNSRecord),
-      order: 'asc',
-      orderBy: 'target',
-      title: 'NS Record',
-    },
-
-    /** MX Record */
-    {
-      columns: [
-        {
-          render: (r: DomainRecord) => r.target,
-          title: 'Mail Server',
-        },
-        {
-          render: (r: DomainRecord) => String(r.priority),
-          title: 'Preference',
-        },
-        {
-          render: (r: DomainRecord) => r.name,
-          title: 'Subdomain',
-        },
-        {
-          render: getTTL,
-          title: 'TTL',
-        },
-        {
-          render: (domainRecordParams: DomainRecord) => {
-            const { id, name, priority, target, ttl_sec } = domainRecordParams;
-            return (
-              <DomainRecordActionMenu
-                deleteData={{
-                  onDelete: confirmDeletion,
-                  recordID: id,
-                }}
-                editPayload={{
-                  id,
-                  name,
-                  priority,
-                  target,
-                  ttl_sec,
-                }}
-                label={name}
-                onEdit={openForEditMXRecord}
-              />
-            );
-          },
-          title: '',
-        },
-      ],
-      data: domainRecords.filter(typeEq('MX')),
-      link: () => createLink('Add a MX Record', openForCreateMXRecord),
-      order: 'asc',
-      orderBy: 'target',
-      title: 'MX Record',
-    },
-
-    /** A/AAAA Record */
-    {
-      columns: [
-        {
-          render: (r: DomainRecord) => r.name || domain.domain,
-          title: 'Hostname',
-        },
-        { render: (r: DomainRecord) => r.target, title: 'IP Address' },
-        { render: getTTL, title: 'TTL' },
-        {
-          render: (domainRecordParams: DomainRecord) => {
-            const { id, name, target, ttl_sec } = domainRecordParams;
-            return (
-              <DomainRecordActionMenu
-                deleteData={{
-                  onDelete: confirmDeletion,
-                  recordID: id,
-                }}
-                editPayload={{
-                  id,
-                  name,
-                  target,
-                  ttl_sec,
-                }}
-                label={name || domain.domain}
-                onEdit={openForEditARecord}
-              />
-            );
-          },
-          title: '',
-        },
-      ],
-      data: domainRecords.filter((r) => typeEq('AAAA', r) || typeEq('A', r)),
-      link: () => createLink('Add an A/AAAA Record', openForCreateARecord),
-      order: 'asc',
-      orderBy: 'name',
-      title: 'A/AAAA Record',
-    },
-
-    /** CNAME Record */
-    {
-      columns: [
-        { render: (r: DomainRecord) => r.name, title: 'Hostname' },
-        { render: (r: DomainRecord) => r.target, title: 'Aliases to' },
-        { render: getTTL, title: 'TTL' },
-        {
-          render: (domainRecordParams: DomainRecord) => {
-            const { id, name, target, ttl_sec } = domainRecordParams;
-            return (
-              <DomainRecordActionMenu
-                deleteData={{
-                  onDelete: confirmDeletion,
-                  recordID: id,
-                }}
-                editPayload={{
-                  id,
-                  name,
-                  target,
-                  ttl_sec,
-                }}
-                label={name}
-                onEdit={openForEditCNAMERecord}
-              />
-            );
-          },
-          title: '',
-        },
-      ],
-      data: domainRecords.filter(typeEq('CNAME')),
-      link: () => createLink('Add a CNAME Record', openForCreateCNAMERecord),
-      order: 'asc',
-      orderBy: 'name',
-      title: 'CNAME Record',
-    },
-
-    /** TXT Record */
-    {
-      columns: [
-        {
-          render: (r: DomainRecord) => r.name || domain.domain,
-          title: 'Hostname',
-        },
-        {
-          render: (r: DomainRecord) => truncateEnd(r.target, 100),
-          title: 'Value',
-        },
-        { render: getTTL, title: 'TTL' },
-        {
-          render: (domainRecordParams: DomainRecord) => {
-            const { id, name, target, ttl_sec } = domainRecordParams;
-            return (
-              <DomainRecordActionMenu
-                deleteData={{
-                  onDelete: confirmDeletion,
-                  recordID: id,
-                }}
-                editPayload={{
-                  id,
-                  name,
-                  target,
-                  ttl_sec,
-                }}
-                label={name}
-                onEdit={openForEditTXTRecord}
-              />
-            );
-          },
-          title: '',
-        },
-      ],
-      data: domainRecords.filter(typeEq('TXT')),
-      link: () => createLink('Add a TXT Record', openForCreateTXTRecord),
-      order: 'asc',
-      orderBy: 'name',
-      title: 'TXT Record',
-    },
-    /** SRV Record */
-    {
-      columns: [
-        { render: (r: DomainRecord) => r.name, title: 'Service/Protocol' },
-        {
-          render: () => domain.domain,
-          title: 'Name',
-        },
-        {
-          render: (r: DomainRecord) => String(r.priority),
-          title: 'Priority',
-        },
-        {
-          render: (r: DomainRecord) => String(r.weight),
-          title: 'Weight',
-        },
-        { render: (r: DomainRecord) => String(r.port), title: 'Port' },
-        { render: (r: DomainRecord) => r.target, title: 'Target' },
-        { render: getTTL, title: 'TTL' },
-        {
-          render: ({
-            id,
-            port,
-            priority,
-            protocol,
-            service,
-            target,
-            weight,
-          }: DomainRecord) => (
-            <DomainRecordActionMenu
-              deleteData={{
-                onDelete: confirmDeletion,
-                recordID: id,
-              }}
-              editPayload={{
-                id,
-                port,
-                priority,
-                protocol,
-                service,
-                target,
-                weight,
-              }}
-              label={domain.domain}
-              onEdit={openForEditSRVRecord}
-            />
-          ),
-          title: '',
-        },
-      ],
-      data: domainRecords.filter(typeEq('SRV')),
-      link: () => createLink('Add an SRV Record', openForCreateSRVRecord),
-      order: 'asc',
-      orderBy: 'name',
-      title: 'SRV Record',
-    },
-
-    /** CAA Record */
-    {
-      columns: [
-        { render: (r: DomainRecord) => r.name, title: 'Name' },
-        { render: (r: DomainRecord) => r.tag, title: 'Tag' },
-        {
-          render: (r: DomainRecord) => r.target,
-          title: 'Value',
-        },
-        { render: getTTL, title: 'TTL' },
-        {
-          render: (domainRecordParams: DomainRecord) => {
-            const { id, name, tag, target, ttl_sec } = domainRecordParams;
-            return (
-              <DomainRecordActionMenu
-                deleteData={{
-                  onDelete: confirmDeletion,
-                  recordID: id,
-                }}
-                editPayload={{
-                  id,
-                  name,
-                  tag,
-                  target,
-                  ttl_sec,
-                }}
-                label={name}
-                onEdit={openForEditCAARecord}
-              />
-            );
-          },
-          title: '',
-        },
-      ],
-      data: domainRecords.filter(typeEq('CAA')),
-      link: () => createLink('Add a CAA Record', openForCreateCAARecord),
-      order: 'asc',
-      orderBy: 'name',
-      title: 'CAA Record',
-    },
-  ];
 
   const defaultDrawerState: DrawerState = {
     mode: 'create',
@@ -543,14 +150,6 @@ export const DomainRecords = (props: Props) => {
       type,
     }));
 
-  const openForCreateARecord = () => openForCreation('AAAA');
-  const openForCreateCAARecord = () => openForCreation('CAA');
-  const openForCreateCNAMERecord = () => openForCreation('CNAME');
-  const openForCreateMXRecord = () => openForCreation('MX');
-  const openForCreateNSRecord = () => openForCreation('NS');
-  const openForCreateSRVRecord = () => openForCreation('SRV');
-  const openForCreateTXTRecord = () => openForCreation('TXT');
-
   const openForEditing = (
     type: DomainType | RecordType,
     fields: Partial<Domain> | Partial<DomainRecord>
@@ -563,34 +162,11 @@ export const DomainRecords = (props: Props) => {
       type,
     }));
 
-  const openForEditARecord = (
-    f: Pick<DomainRecord, 'id' | 'name' | 'target' | 'ttl_sec'>
-  ) => openForEditing('AAAA', f);
-  const openForEditCAARecord = (
-    f: Pick<DomainRecord, 'id' | 'name' | 'tag' | 'target' | 'ttl_sec'>
-  ) => openForEditing('CAA', f);
-  const openForEditCNAMERecord = (
-    f: Pick<DomainRecord, 'id' | 'name' | 'target' | 'ttl_sec'>
-  ) => openForEditing('CNAME', f);
-  const openForEditMXRecord = (
-    f: Pick<DomainRecord, 'id' | 'name' | 'priority' | 'target' | 'ttl_sec'>
-  ) => openForEditing('MX', f);
-  const openForEditNSRecord = (
-    f: Pick<DomainRecord, 'id' | 'name' | 'target' | 'ttl_sec'>
-  ) => openForEditing('NS', f);
   const openForEditPrimaryDomain = (f: Partial<Domain>) =>
     openForEditing('master', f);
-  const openForEditSRVRecord = (
-    f: Pick<
-      DomainRecord,
-      'id' | 'name' | 'port' | 'priority' | 'protocol' | 'target' | 'weight'
-    >
-  ) => openForEditing('SRV', f);
+
   const openForEditSecondaryDomain = (f: Partial<Domain>) =>
     openForEditing('slave', f);
-  const openForEditTXTRecord = (
-    f: Pick<DomainRecord, 'id' | 'name' | 'target' | 'ttl_sec'>
-  ) => openForEditing('TXT', f);
 
   const renderDialogActions = () => {
     return (
@@ -627,12 +203,31 @@ export const DomainRecords = (props: Props) => {
     });
   };
 
+  const handlers: GenerateTypesHandlers = {
+    confirmDeletion,
+    handleOpenSOADrawer,
+    openForCreateARecord: () => openForCreation('AAAA'),
+    openForCreateCAARecord: () => openForCreation('CAA'),
+    openForCreateCNAMERecord: () => openForCreation('CNAME'),
+    openForCreateMXRecord: () => openForCreation('MX'),
+    openForCreateNSRecord: () => openForCreation('NS'),
+    openForCreateSRVRecord: () => openForCreation('SRV'),
+    openForCreateTXTRecord: () => openForCreation('TXT'),
+    openForEditARecord: (f) => openForEditing('AAAA', f),
+    openForEditCAARecord: (f) => openForEditing('CAA', f),
+    openForEditCNAMERecord: (f) => openForEditing('CNAME', f),
+    openForEditMXRecord: (f) => openForEditing('MX', f),
+    openForEditNSRecord: (f) => openForEditing('NS', f),
+    openForEditSRVRecord: (f) => openForEditing('SRV', f),
+    openForEditTXTRecord: (f) => openForEditing('TXT', f),
+  };
+
   React.useEffect(() => {
     setState((prevState) => ({
       ...prevState,
-      types: generateTypes(),
+      types: generateTypes(props, handlers),
     }));
-  }, [domainRecords, domain]);
+  }, [domain, domainRecords]);
 
   return (
     <>
